Combine into one eventprocessor
