+++ Source/Analytics/Core/.dolittle/artifacts.json
@@ -7,6 +7,14 @@
       "893b09bd-5d69-4a05-bc2a-420ba95cd169": {
         "generation": 1,
         "type": "Read.CaseReports.CaseReport, Read"
+      },
+      "2045add7-bc9d-4abb-85d1-c2b23e20d920": {
+        "generation": 1,
+        "type": "Read.CaseReports.CaseReportsInRegionLast7Days, Read"
+      },
+      "01499494-3a94-4333-9e63-f5c63d10c59f": {
+        "generation": 1,
+        "type": "Read.CaseReports.CaseReportsPerRegionLast7Days, Read"
       }
     },
     "queries": {
