+++ Source/Analytics/Web/src/Features/CaseReports/CaseReportsInRegionLast7Days.js
@@ -1,0 +1,19 @@
+﻿/*---------------------------------------------------------------------------------------------
+ *  This file is an automatically generated ReadModel Proxy
+ *  
+ *--------------------------------------------------------------------------------------------*/
+import { ReadModel } from  '@dolittle/readmodels';
+
+export class CaseReportsInRegionLast7Days extends ReadModel
+{
+    constructor() {
+        super();
+        this.artifact = {
+           id: '2045add7-bc9d-4abb-85d1-c2b23e20d920',
+           generation: '1'
+        };
+        this.id = 0;
+        this.region = '';
+        this.numberOfCaseReports = 0;
+    }
+}
