+++ Source/Analytics/Web/src/Features/CaseReports/CaseReportsPerRegionLast7Days.js
@@ -1,0 +1,18 @@
+﻿/*---------------------------------------------------------------------------------------------
+ *  This file is an automatically generated ReadModel Proxy
+ *  
+ *--------------------------------------------------------------------------------------------*/
+import { ReadModel } from  '@dolittle/readmodels';
+
+export class CaseReportsPerRegionLast7Days extends ReadModel
+{
+    constructor() {
+        super();
+        this.artifact = {
+           id: '01499494-3a94-4333-9e63-f5c63d10c59f',
+           generation: '1'
+        };
+        this.id = 0;
+        this.caseReportsPerRegion = [];
+    }
+}
