+++ Source/Analytics/Web/src/Features/CaseReports/PerRegion/CaseReportsInRegionLast7Days.js
@@ -1,0 +1,19 @@
+﻿/*---------------------------------------------------------------------------------------------
+ *  This file is an automatically generated ReadModel Proxy
+ *  
+ *--------------------------------------------------------------------------------------------*/
+import { ReadModel } from  '@dolittle/readmodels';
+
+export class CaseReportsInRegionLast7Days extends ReadModel
+{
+    constructor() {
+        super();
+        this.artifact = {
+           id: '8ddd24d4-0d1f-4076-a65c-849e15a9ee05',
+           generation: '1'
+        };
+        this.id = 0;
+        this.region = '';
+        this.numberOfCaseReports = 0;
+    }
+}
