+++ Source/Analytics/Web/src/Features/CaseReports/PerRegion/CaseReportsPerRegionLast7Days.js
@@ -1,0 +1,18 @@
+﻿/*---------------------------------------------------------------------------------------------
+ *  This file is an automatically generated ReadModel Proxy
+ *  
+ *--------------------------------------------------------------------------------------------*/
+import { ReadModel } from  '@dolittle/readmodels';
+
+export class CaseReportsPerRegionLast7Days extends ReadModel
+{
+    constructor() {
+        super();
+        this.artifact = {
+           id: '0db03ac7-cd74-462c-8fb1-1f683fd84b8e',
+           generation: '1'
+        };
+        this.id = 0;
+        this.caseReportsPerRegion = [];
+    }
+}
